% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InSilicoVA2020-package.R
\docType{package}
\name{InSilicoVA2020-package}
\alias{InSilicoVA2020-package}
\title{Probabilistic Verbal Autopsy Coding with 'InSilicoVA' Algorithm and the WHO 2020 VA Instrument}
\description{
Computes individual causes of death and population cause-specific mortality
fractions using the 'InSilicoVA' algorithm from McCormick et al. (2016)
<DOI:10.1080/01621459.2016.1152191>. It uses data derived from the WHO 2020
VA instrument. This package provides general model fitting and customization
for 'InSilicoVA' algorithm and basic graphical visualization of the output.
}
\references{
http://openva.net
}
\author{
Zehang Li, Tyler McCormick, Peter Chio, Jason Thomas, Yue Chu, Sam Clark

Maintainer: Jason Thomas <jarathomas@gmail.com>
}
\keyword{InSilico}
