% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data_2016}
\alias{prep_data_2016}
\title{Prepare data collected with WHO VA 2016 instrument}
\usage{
prep_data_2016(data)
}
\arguments{
\item{data}{Input data collected using the WHO VA 2016 instrument.}
}
\value{
Data frame with the format expected by InSilicoVA2020.
}
\description{
This function takes data collected using the WHO VA 2016 instrument, and
converts it to the format that is consistent with the WHO VA 2020 instrument.
The conversion consists of setting the values for the missing indicators/columns
to missing
}
