% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data_2020}
\alias{prep_data_2020}
\title{Prepare data collected with the WHO VA 2020 instrument}
\usage{
prep_data_2020(data)
}
\arguments{
\item{data}{Input data collected with the WHO VA 2020 insturment.}
}
\value{
Data frame with the format expected by InSilicoVA2020.
}
\description{
This function takes data collected using the WHO VA 2020 instrument, and
converts it to the format expected by InSilicoVA. The conversion consists
of adding new indicators/columns (with missing values).
}
