% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InSilicoVA2020-package.R
\docType{data}
\name{probbaseV5_2020}
\alias{probbaseV5_2020}
\title{Conditional probability of InterVA5}
\format{
A data frame with 354 observations on 87 variables. The first row contains
observations corresponding to prior probabilities; while the subsequent observations
(rows 2 - 354) are the conditional probabilities.
}
\description{
This is the table of conditional probabilities of symptoms given CODs, along with
prior probabilities in the first row. The values are from InterVA2020 and correspond
to the WHO VA 2016 instrument, but modified to the WHO VA 2020 instrument format.
}
\examples{

data(probbaseV5_2020)

}
\keyword{datasets}
