% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InSilicoVA2020-package.R
\docType{data}
\name{RandomVA2020}
\alias{RandomVA2020}
\title{200 records of Sample Input}
\format{
200 arbitrary input records.
}
\description{
This is a dataset consisting of 200 arbitrary sample input deaths in the
acceptable format of InterVA2020. Any dataset that needs to be analyzed by this
package should be in the same format. The order of the input fields must
not be changed.
}
\examples{

data(RandomVA2020)

}
\keyword{datasets}
